+++ packages/validator/index.ts
@@ -1,2 +1,3 @@
 export * from './src/email/index';
 export * from './src/mobile/index';
+export * from './src/citizenID/index';